Fix exponential lookup, comparisons
I found lookupVx was taking up 97% of the runtime on large files. This was because it always did a direct search of the cross table and the value table.

Instead, get the gid using the value table's hashtable, and do a linear search of the cross table for the resulting xid.

Also remove byte-by-byte comparisons on strings. Favor strncmp.